requirements.txt
fastapi==0.104.1
pydantic==2.5.0
sqlmodel==0.0.16
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
requests==2.31.0
